+++ src/content/github/dom.ts
@@ -1,0 +1,27 @@
+import { extractSlugFromCurrentUrl, setGitHubUser } from "@r2c/extension/utils";
+
+export async function extractCurrentUserFromPage(): Promise<
+  string | undefined
+> {
+  const { domain } = extractSlugFromCurrentUrl();
+
+  if (domain.includes("github.com")) {
+    const userLoginMetaTags = document.getElementsByName("user-login");
+
+    if (userLoginMetaTags.length === 0) {
+      return undefined;
+    }
+
+    const user = userLoginMetaTags[0].getAttribute("content");
+
+    if (user != null && user !== "") {
+      setGitHubUser(user);
+
+      return user;
+    } else {
+      return undefined;
+    }
+  } else {
+    return undefined;
+  }
+}
